add option to randomize the resource filename (#76)
* fix 32bit types
* add argument to control the randomizing of the filename

+++ examples/tf_inference.c
@@ -72,7 +72,7 @@ int main(int argc, char *argv[])
 	/* Input tensors & nodes */
 	struct vaccel_tf_node in_node = { "serving_default_input_1", 0 };
 	
-	long long int dims[] = {1, 30};
+	uint32_t dims[] = {1, 30};
 	float data[30];
 	for (int i = 0; i < 30; ++i)
 		data[i] = 1.00;
@@ -102,7 +102,7 @@ int main(int argc, char *argv[])
 	printf("Success!\n");
 	printf("Output tensor => type:%u nr_dims:%u\n", out->data_type, out->nr_dims);
 	for (int i = 0; i < out->nr_dims; ++i)
-		printf("dim[%d]: %lld\n", i, out->dims[i]);
+		printf("dim[%d]: %d\n", i, out->dims[i]);
 		
 	printf("Result Tensor :\n");
 	float *offsets = (float *)out->data;
