+++ src/include/ops/tf.h
@@ -93,7 +93,7 @@ struct vaccel_tf_tensor {
 
         /* dimensions of the data */
         int nr_dims;
-        long long int *dims;
+        uint32_t *dims;
 
         /* Data type */
         enum vaccel_tf_data_type data_type;
@@ -102,13 +102,13 @@ struct vaccel_tf_tensor {
 struct vaccel_tf_tensor *
 vaccel_tf_tensor_new(
 	int nr_dims,
-	long long int *dims,
+	uint32_t  *dims,
 	enum vaccel_tf_data_type type
 );
 
 struct vaccel_tf_tensor *
 vaccel_tf_tensor_allocate(
-	int nr_dims, long long int *dims,
+	int nr_dims, uint32_t *dims,
 	enum vaccel_tf_data_type type,
 	size_t total_size
 );
